# public/openapi.yaml
openapi: 3.0.2
info:
  version: 1.0.0
  title: Businesses API
  description: A simple API for a Yelp-like application.
paths:
...
  '/businesses/{businessid}':
    parameters:
      - name: businessid
        in: path
        description: Unique ID of a business.
        required: true
        schema:
          type: string
          example: 5ce48a2ddf60d448aed2b1d5
    get:
      summary: Fetch data for a specific business.
      description: >
        Returns complete data for a the business specified by `businessid`.
      operationId: getBusinessById
      tags:
        - Businesses
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Business'
        '404':
          description: Specified `businessid` not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update data for a specific business.
      description: |
        Replaces the data for the business specified by `businessid`.
      operationId: replaceBusiness
      tags:
        - Businesses
      requestBody:
        description: |
          Business data to replace data for the business specified by `businessid`.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Business'
      responses:
        '200':
          description: Success
        '400':
          description: The request body was either not present or did not contain all of the required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Specified `businessid` not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Remove a specific business from the database.
      description: |
        Completely removes the data for the business specified by `businessid`.
      operationId: removeBusiness
      tags:
        - Businesses
      responses:
        '204':
          description: Success
        '404':
          description: Specified `businessid` not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /reviews:
    post:
      summary: Add a new review
      description: >
        Creates a new review with specified data and adds it to the
        application's database.
      operationId: addNewReview
      tags:
        - Reviews
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
      responses:
        '201':
          description: New review successfully added
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: Unique ID of the created review.
                    example: 10
        '400':
          description: >
            The request body was either not present or did not contain all of the required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: >
            The specified user has already posted a review for the specified business.  A given user may only post one review for a given business.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/reviews/{reviewID}':
    parameters:
      - name: reviewID
        in: path
        description: Unique ID of a review.
        schema:
          type: integer
        example: 0
        required: true
    get:
      summary: Fetch data for a specific review.
      description: >
        Returns complete data for a the review specified by `reviewID`.
      operationId: getReview
      tags:
        - Reviews
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '404':
          description: Specified `reviewID` not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      summary: Update data for a specific review.
      description: >
        Replaces the data for the review specified by `review`.
      operationId: replaceReview
      tags:
        - Reviews
      requestBody:
        description: >
          Review data to replace data for the review specified by `reviewID`.
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
      responses:
        '200':
          description: Success
        '400':
          description: >
            The request body was either not present or did not contain all of the required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: >
            The request tried to change the `userid` or the `businessid` for a review.  Once a review is created, these cannot be changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Specified `reviewID` not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Remove a specific review from the database.
      description: >
        Completely removes the data for the review specified by `reviewID`.
      operationId: removeReview
      tags:
        - Reviews
      responses:
        '204':
          description: Success
        '404':
          description: Specified `reviewID` not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users:
    post:
      summary: Add a new application user
      description: >
        Creates a new user with specified data and adds it to the
        application's database.
      operationId: addNewUser
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        '201':
          description: New user successfully added
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                    description: Unique ID of the created user.
                    example: 19
        '400':
          description: The request body was either not present or did not contain all of the required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/login:
    post:
      summary: Log in a User.
      description: >
        Authenticate a specific User with their email address and password.
      operationId: authenticateUser
      tags:
        - Users
      requestBody:
        required: true
        description: >
          Email address and plain-text password for the user being authenticated.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
            example:
              email: TheLastJedi@gmail.com
              password: hunter2
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: >
                      A JWT authentication token.
                    example: aaaaaaaa.bbbbbbbb.cccccccc

        '400':
          description: >
            The request body was either not present or did not contain all of the required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: >
            The specified credentials were invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: >
            An internal server error occurred.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  '/users/{userID}':
    parameters:
      - name: userID
        in: path
        description: Unique ID of a user.
        schema:
          type: integer
        example: 5
        required: true
    get:
      summary: Fetch data for a specific user.
      description: >
        Returns complete data for a the user specified by `userID`.
      operationId: getUser
      tags:
        - Users
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: Specified `userID` not found.  This may be returned if a user is not authorized to access the requested data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /users/{userid}/businesses:
    parameters:
      - name: userid
        in: path
        description: Unique ID of a user.
        schema:
          type: integer
        example: 0
        required: true
    get:
      summary: Fetch a user's owned businesses.
      description: >
        Returns a list of all of the businesses owned by the user specified by `userid`.
      operationId: getUserBusinesses
      tags:
        - Users
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  lodgings:
                    type: array
                    description: The businesses owned by the specified user.
                    items:
                      $ref: '#/components/schemas/Business'
  /users/{userid}/reviews:
    parameters:
      - name: userid
        in: path
        description: Unique ID of a user.
        schema:
          type: integer
        example: 21
        required: true
    get:
      summary: Fetch a user's reviews.
      description: >
        Returns a list of all of the reviews written by the user specified by `userid`.
      operationId: getUserReviews
      tags:
        - Users
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  lodgings:
                    type: array
                    description: The reviews written by the specified user.
                    items:
                      $ref: '#/components/schemas/Review'
  /users/{userid}/photos:
    parameters:
      - name: userid
        in: path
        description: Unique ID of a user.
        schema:
          type: integer
        example: 21
        required: true
    get:
      summary: Fetch a user's photos.
      description: >
        Returns a list of all of the photos posted by the user specified by `userid`.
      operationId: getUserPhotos
      tags:
        - Users
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  lodgings:
                    type: array
                    description: The photos posted by the specified user.
                    items:
                      $ref: '#/components/schemas/Photo'
components:
  schemas:
    User:
      description: >
        An object representing information about an application user.
      type: object
      properties:
        name:
          type: string
          description: Real name of the user.  This is required when creating a new User.
          example: Luke Skywalker
        email:
          type: string
          description: The user's email address.  This is required to be unique among all Users.
          example: TheLastJedi@gmail.com
        password:
          type: string
          description: The user's plaintext password.  This is required when creating a new User and when logging in.
          example: hunter2
        admin:
          type: boolean
          description: Boolean flag indicating whether the user has admin permissions.
          default: false
      required:
        - name
        - email
        - password
    Business:
      description: >
        An object representing information about a single business.
      type: object
      properties:
        name:
          type: string
          description: Name of the business.
          example: American Dream Pizza
        address:
          type: object
          properties:
            street:
              type: string
              description: Street address of the business.
              example: 2525 NW Monroe Ave.
            city:
              type: string
              description: The city in which the business resides.
              example: Corvallis
            state:
              type: string
              description: The state in which the business resides.
              example: OR
            zip:
              type: string
              description: >
                The zip code of the city in which the business resides.
              example: '97330'
        phone:
          type: string
          description: The business' phone number.
          example: '541-757-1713'
        category:
          type: string
          description: The business' category.
          example: 'Restaurant'
        subcategory:
          type: string
          description: The business' subcategory.
          example: 'Pizza'
        website:
          type: string
          description: The business' website URL.
          example: 'http://adpizza.com'
        email:
          type: string
          description: The business' email address.
          example: 'brian@adpizza.com'
        photos:
          type: array
          description: >
            An array of photos of the business.  These are only returned when
            fetching detailed information about a business.
          items:
            $ref: '#/components/schemas/Photo'
      required:
        - name
        - address
        - phone
        - category
        - subcategory
    Photo:
      description: >
        An object representing information about a photo of a business.
      type: object
      properties:
        businessid:
          type: string
          description: >
            The ID of the business associated with this photo.  This is
            required when creating a photo.
          example: 5ce48a2ddf60d448aed2b1d5
        caption:
          type: string
          description: Caption text for the photo
          example: "Kale!"
        file:
          type: string
          description: >
            When the photo is being uploaded, this field will hold the binary
            data contained in the photo file.  This is required only when
            creating a photo.
        urls:
          type: object
          description: >
            When photo information is fetched from the API, this field will
            contain an object that specifies all images available for
            download for this photo.  The object will contain a single entry
            per available image size.  For example, if the '1024' image size
            is available for this photo, this object will contain a '1024'
            entry.
          example:
            '1024': /media/photos/5ce48a2ddf60d448aed2b1c3-1024.jpg
            '640': /media/photos/5ce48a2ddf60d448aed2b1c3-640.jpg
            '256': /media/photos/5ce48a2ddf60d448aed2b1c3-256.jpg
            '128': /media/photos/5ce48a2ddf60d448aed2b1c3-128.jpg
            'orig': /media/photos/5ce48a2ddf60d448aed2b1c3-orig.jpg
      required:
        - businessid
        - file
    Review:
      description: >
        An object representing information about a review of a business.
      type: object
      properties:
        userid:
          type: integer
          description: >
            The ID of the user associated with this review.  This is required when creating or updating a review.
          example: 7
        businessid:
          type: integer
          description: >
            The ID of the business associated with this review.  This is required when creating or updating a review.
          example: 3
        dollars:
          type: integer
          description: A rating of the expensiveness of a business.  Must be an integer between 1 and 4.
          example: 3
          minimum: 1
          maximum: 4
        stars:
          type: number
          description: A rating of the quality of a business.  Must be an number between 0 and 5.
          example: 5
          minimum: 0
          maximum: 5
        review:
          type: string
          description: Textual business review.
          example: "Great produce!"
      required:
        - userid
        - businessid
        - dollars
        - stars
    Error:
      description: >
        An object representing an error response from the API.
      type: object
      properties:
        error:
          type: string
          description: A message describing the error.
...